ABW-2806 - NPSSurveyApi moved to separate package.
