+++ app/src/main/java/com/babylon/wallet/android/di/NetworkModule.kt
@@ -3,7 +3,7 @@ package com.babylon.wallet.android.di
 import com.babylon.wallet.android.BuildConfig
 import com.babylon.wallet.android.data.dapp.PeerdroidClient
 import com.babylon.wallet.android.data.dapp.PeerdroidClientImpl
-import com.babylon.wallet.android.data.gateway.apis.NPSSurveyApi
+import com.babylon.wallet.android.data.gateway.survey.NPSSurveyApi
 import com.babylon.wallet.android.data.gateway.apis.StateApi
 import com.babylon.wallet.android.data.gateway.apis.StreamApi
 import com.babylon.wallet.android.data.gateway.apis.TransactionApi
